refactor: 상품 mapper 테스트 일부를 integration test로 이동
