+++ src/test/groovy/kr/flab/fream/IntegrationSpec.groovy
@@ -1,0 +1,14 @@
+package kr.flab.fream
+
+import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase
+import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.context.annotation.Import
+import org.springframework.transaction.annotation.Transactional
+
+
+@SpringBootTest
+@Import(DatabaseClearConfig)
+@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
+@Transactional
+class IntegrationSpec extends DatabaseTest {
+}
